去除编译告警（未使用的变量和类型匹配）
git-svn-id: http://192.168.20.11/sequoiadb/trunk@18977 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/cls/clsDCMgr.cpp
@@ -696,7 +696,8 @@ namespace engine
       hasLock = TRUE ;
 
       // first send to primary node
-      if ( _peerCatPrimary >= 0 && _peerCatPrimary < _vecCatlog.size() )
+      if ( _peerCatPrimary >= 0 && 
+           ( UINT32 )_peerCatPrimary < _vecCatlog.size() )
       {
          rc = _syncSend2PeerNode( cb, msg, ppRecvMsg,
                                   _vecCatlog[ _peerCatPrimary ],
